+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # 5.3.1
 - add config comment for StopRenderingSkippedEntities option
-- Add EntityTickableRefreshInterval config option (with comment)
+- Add EntityRenderableRefreshInterval config option (with comment)